requirements.txt
langchain
langchain-community
langchain-core
unstructured
sec-api
python-dotenv
crewai
faiss-cpu
torch
torchvision
torchaudio
tools
tiktoken
langchain_groq
streamlit